better handled the functions associated to namespace and reconciler
