+++ operators/pkg/controller/tenant/reconciler.go
@@ -28,12 +28,20 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/util/workqueue"
 	"k8s.io/klog/v2"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/utils/ptr"
+
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/source"
+	"github.com/netgroup-polito/CrownLabs/operators/pkg/forge"
+
 
 	crownlabsv1alpha2 "github.com/netgroup-polito/CrownLabs/operators/api/v1alpha2"
 	"github.com/netgroup-polito/CrownLabs/operators/pkg/controller/utils"
@@ -88,7 +96,14 @@ func (r *TenantReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctr
 		return ctrl.Result{}, client.IgnoreNotFound(err)
 	}
 
+	if tn.Labels[r.TargetLabelKey] != r.TargetLabelValue {
+		// if entered here it means that is in the reconcile
+		// which has been requed after
+		// the last successful one with the old target label
+		return ctrl.Result{}, nil
+	}
 
+	
 	// Determine if the personal namespace should be deleted
 
 	nsName := fmt.Sprintf("tenant-%s", strings.ReplaceAll(tn.Name, ".", "-"))
@@ -180,4 +195,176 @@ func (r *TenantReconciler) SetupWithManager(mgr ctrl.Manager) error {
 		// }).
 		// WithLogConstructor(utils.LogConstructor(mgr.GetLogger(), "Tenant")).
 		Complete(r)
+}
+
+func (r *TenantReconciler) createOrUpdateTnPersonalNFSVolume(ctx context.Context, tn *crownlabsv1alpha2.Tenant, nsName string) error {
+	// Persistent volume claim NFS
+	pvc := v1.PersistentVolumeClaim{ObjectMeta: metav1.ObjectMeta{Name: myDrivePVCName(tn.Name), Namespace: r.MyDrivePVCsNamespace}}
+
+	pvcOpRes, err := ctrl.CreateOrUpdate(ctx, r.Client, &pvc, func() error {
+		r.updateTnPersistentVolumeClaim(&pvc)
+		return ctrl.SetControllerReference(tn, &pvc, r.Scheme)
+	})
+	if err != nil {
+		klog.Errorf("Unable to create or update PVC for tenant %s -> %s", tn.Name, err)
+		return err
+	}
+	klog.Infof("PVC for tenant %s %s", tn.Name, pvcOpRes)
+
+	if pvc.Status.Phase == v1.ClaimBound {
+		pv := v1.PersistentVolume{ObjectMeta: metav1.ObjectMeta{Name: pvc.Spec.VolumeName}}
+		if err := r.Get(ctx, types.NamespacedName{Name: pv.Name}, &pv); err != nil {
+			klog.Errorf("Unable to get PV for tenant %s -> %s", tn.Name, err)
+			return err
+		}
+		pvcSecret := v1.Secret{ObjectMeta: metav1.ObjectMeta{Name: NFSSecretName, Namespace: nsName}}
+		pvcSecOpRes, err := ctrl.CreateOrUpdate(ctx, r.Client, &pvcSecret, func() error {
+			r.updateTnPVCSecret(&pvcSecret, fmt.Sprintf("%s.%s", pv.Spec.CSI.VolumeAttributes["server"], pv.Spec.CSI.VolumeAttributes["clusterID"]), pv.Spec.CSI.VolumeAttributes["share"])
+			return ctrl.SetControllerReference(tn, &pvcSecret, r.Scheme)
+		})
+		if err != nil {
+			klog.Errorf("Unable to create or update PVC Secret for tenant %s -> %s", tn.Name, err)
+			return err
+		}
+		klog.Infof("PVC Secret for tenant %s %s", tn.Name, pvcSecOpRes)
+
+		val, found := pvc.Labels[forge.ProvisionJobLabel]
+		if !found || val != forge.ProvisionJobValueOk {
+			chownJob := batchv1.Job{ObjectMeta: metav1.ObjectMeta{Name: pvc.Name + "-provision", Namespace: pvc.Namespace}}
+			labelToSet := forge.ProvisionJobValuePending
+
+			chownJobOpRes, err := ctrl.CreateOrUpdate(ctx, r.Client, &chownJob, func() error {
+				if chownJob.CreationTimestamp.IsZero() {
+					klog.Infof("PVC Provisioning Job created for tenant %s", tn.Name)
+					r.updateTnProvisioningJob(&chownJob, &pvc)
+				} else if found && val == forge.ProvisionJobValuePending {
+					if chownJob.Status.Succeeded == 1 {
+						labelToSet = forge.ProvisionJobValueOk
+						klog.Infof("PVC Provisioning Job completed for tenant %s", tn.Name)
+					} else if chownJob.Status.Failed == 1 {
+						klog.Warningf("PVC Provisioning Job failed for tenant %s", tn.Name)
+					}
+				}
+
+				return ctrl.SetControllerReference(tn, &chownJob, r.Scheme)
+			})
+			if err != nil {
+				klog.Errorf("Unable to create or update PVC Provisioning Job for tenant %s -> %s", tn.Name, err)
+				return err
+			}
+			klog.Infof("PVC Provisioning Job for tenant %s %s", tn.Name, chownJobOpRes)
+
+			pvc.Labels[forge.ProvisionJobLabel] = labelToSet
+			if err := r.Update(ctx, &pvc); err != nil {
+				klog.Errorf("PVC Provisioning Job failed to update PVC labels for tenant %s", tn.Name)
+			}
+			klog.Infof("PVC Provisioning Job updateded PVC label to %s for tenant %s", labelToSet, tn.Name)
+		}
+	} else if pvc.Status.Phase == v1.ClaimPending {
+		klog.Infof("PVC pending for tenant %s", tn.Name)
+	}
+	return nil
+}
+
+func (r *TenantReconciler) updateTnRb(rb *rbacv1.RoleBinding, tnName string) {
+	rb.Labels = r.updateTnResourceCommonLabels(rb.Labels)
+	rb.RoleRef = rbacv1.RoleRef{Kind: "ClusterRole", Name: "crownlabs-manage-instances", APIGroup: "rbac.authorization.k8s.io"}
+	rb.Subjects = []rbacv1.Subject{{Kind: "User", Name: tnName, APIGroup: "rbac.authorization.k8s.io"}}
+}
+
+func (r *TenantReconciler) updateTnCr(rb *rbacv1.ClusterRole, tnName string) {
+	rb.Labels = r.updateTnResourceCommonLabels(rb.Labels)
+	rb.Rules = []rbacv1.PolicyRule{{
+		APIGroups:     []string{"crownlabs.polito.it"},
+		Resources:     []string{"tenants"},
+		ResourceNames: []string{tnName},
+		Verbs:         []string{"get", "list", "watch", "patch", "update"},
+	}}
+}
+
+func (r *TenantReconciler) updateTnCrb(rb *rbacv1.ClusterRoleBinding, tnName, crName string) {
+	rb.Labels = r.updateTnResourceCommonLabels(rb.Labels)
+	rb.RoleRef = rbacv1.RoleRef{Kind: "ClusterRole", Name: crName, APIGroup: "rbac.authorization.k8s.io"}
+	rb.Subjects = []rbacv1.Subject{{Kind: "User", Name: tnName, APIGroup: "rbac.authorization.k8s.io"}}
+}
+
+func (r *TenantReconciler) updateTnProvisioningJob(chownJob *batchv1.Job, pvc *v1.PersistentVolumeClaim) {
+	if chownJob.CreationTimestamp.IsZero() {
+		chownJob.Spec.BackoffLimit = ptr.To[int32](ProvisionJobMaxRetries)
+		chownJob.Spec.TTLSecondsAfterFinished = ptr.To[int32](ProvisionJobTTLSeconds)
+		chownJob.Spec.Template.Spec.RestartPolicy = v1.RestartPolicyOnFailure
+		chownJob.Spec.Template.Spec.Containers = []v1.Container{{
+			Name:    "chown-container",
+			Image:   ProvisionJobBaseImage,
+			Command: []string{"chown", "-R", fmt.Sprintf("%d:%d", forge.CrownLabsUserID, forge.CrownLabsUserID), forge.MyDriveVolumeMountPath},
+			VolumeMounts: []v1.VolumeMount{{
+				Name:      "mydrive",
+				MountPath: forge.MyDriveVolumeMountPath,
+			},
+			},
+			Resources: v1.ResourceRequirements{
+				Requests: v1.ResourceList{
+					"cpu":    resource.MustParse("100m"),
+					"memory": resource.MustParse("128Mi"),
+				},
+				Limits: v1.ResourceList{
+					"cpu":    resource.MustParse("100m"),
+					"memory": resource.MustParse("128Mi"),
+				},
+			},
+		},
+		}
+		chownJob.Spec.Template.Spec.Volumes = []v1.Volume{{
+			Name: "mydrive",
+			VolumeSource: v1.VolumeSource{
+				PersistentVolumeClaim: &v1.PersistentVolumeClaimVolumeSource{
+					ClaimName: pvc.Name,
+				},
+			},
+		},
+		}
+	}
+}
+
+
+func (r *TenantReconciler) updateTnPVCSecret(sec *v1.Secret, dnsName, path string) {
+	sec.Labels = r.updateTnResourceCommonLabels(sec.Labels)
+
+	sec.Type = v1.SecretTypeOpaque
+	sec.Data = make(map[string][]byte, 2)
+	sec.Data[NFSSecretServerNameKey] = []byte(dnsName)
+	sec.Data[NFSSecretPathKey] = []byte(path)
+}
+
+
+func (r *TenantReconciler) updateTnNetPolDeny(np *netv1.NetworkPolicy) {
+	np.Labels = r.updateTnResourceCommonLabels(np.Labels)
+	np.Spec.PodSelector.MatchLabels = make(map[string]string)
+	np.Spec.Ingress = []netv1.NetworkPolicyIngressRule{{From: []netv1.NetworkPolicyPeer{{PodSelector: &metav1.LabelSelector{}}}}}
+}
+
+
+func (r *TenantReconciler) updateTnNetPolAllow(np *netv1.NetworkPolicy) {
+	np.Labels = r.updateTnResourceCommonLabels(np.Labels)
+	np.Spec.PodSelector.MatchLabels = make(map[string]string)
+	np.Spec.Ingress = []netv1.NetworkPolicyIngressRule{{From: []netv1.NetworkPolicyPeer{{NamespaceSelector: &metav1.LabelSelector{
+		MatchLabels: map[string]string{"crownlabs.polito.it/allow-instance-access": "true"},
+	}}}}}
+}
+
+func (r *TenantReconciler) updateTnPersistentVolumeClaim(pvc *v1.PersistentVolumeClaim) {
+	scName := r.MyDrivePVCsStorageClassName
+	pvc.Labels = r.updateTnResourceCommonLabels(pvc.Labels)
+
+	pvc.Spec.AccessModes = []v1.PersistentVolumeAccessMode{v1.ReadWriteMany}
+	pvc.Spec.StorageClassName = &scName
+
+	oldSize := *pvc.Spec.Resources.Requests.Storage()
+	if sizeDiff := r.MyDrivePVCsSize.Cmp(oldSize); sizeDiff > 0 || oldSize.IsZero() {
+		pvc.Spec.Resources.Requests = v1.ResourceList{v1.ResourceStorage: r.MyDrivePVCsSize}
+	}
+}
+
+func myDrivePVCName(tnName string) string {
+	return fmt.Sprintf("%s-drive", strings.ReplaceAll(tnName, ".", "-"))
 }
